feat: upgrade immich/ui
