+++ web/src/lib/constants.ts
@@ -78,12 +78,6 @@ export const timeBeforeShowLoadingSpinner: number = 100;
 
 export const timeDebounceOnSearch: number = 300;
 
-// should be the same values as the ones in the app.html
-export enum Theme {
-  LIGHT = 'light',
-  DARK = 'dark',
-}
-
 export const fallbackLocale = {
   code: 'en-US',
   name: 'English (US)',
