+++ web/src/lib/stores/preferences.store.ts
@@ -1,12 +1,7 @@
 import { browser } from '$app/environment';
-import { Theme, defaultLang } from '$lib/constants';
+import { defaultLang } from '$lib/constants';
 import { getPreferredLocale } from '$lib/utils/i18n';
 import { persisted } from 'svelte-persisted-store';
-
-export interface ThemeSetting {
-  value: Theme;
-  system: boolean;
-}
 
 // Locale to use for formatting dates, numbers, etc.
 export const locale = persisted('locale', 'default', {
